+++ PokerTracker/Model/DetailedStatistics/DetailedStatistics.swift
@@ -149,7 +149,7 @@ extension DetailedStatistics: Equatable
         return (
             lhs.id_player == rhs.id_player &&
             lhs.VPIP.value == rhs.VPIP.value &&
-            lhs.PFR.value == rhs.PFR.value
+            lhs.aligned.PFR.value == rhs.aligned.PFR.value
         )
     }
 }
@@ -161,7 +161,7 @@ extension DetailedStatistics: CustomStringConvertible
     
     public var description: String
     {
-        return "id_player: \(id_player), id_site: \(id_site), str_player_name: \(str_player_name), cnt_vpip: \(cnt_vpip), cnt_hands: \(cnt_hands), cnt_walks: \(cnt_walks), cnt_pfr: \(cnt_pfr), cnt_pfr_opp: \(cnt_pfr_opp) \n VPIP:\(String(format: "%.2f%%", VPIP.value)), PFR:\(String(format: "%.2f%%", PFR.value))"
+        return "id_player: \(id_player), id_site: \(id_site), str_player_name: \(str_player_name), cnt_vpip: \(cnt_vpip), cnt_hands: \(cnt_hands), cnt_walks: \(cnt_walks), cnt_pfr: \(cnt_pfr), cnt_pfr_opp: \(cnt_pfr_opp) \n VPIP:\(String(format: "%.2f%%", VPIP.value)), aligned.PFR:\(String(format: "%.2f%%", aligned.PFR.value))"
     }
 }
 
@@ -172,27 +172,35 @@ extension DetailedStatistics
 {
     
     
-    public struct Statistic
+    public struct Statistic: CustomStringConvertible
     {
         
         
+        public let name: String
         public let value: Double
         public let count: Int
         public let opportunities: Int
+        
+        public var description: String
+        {
+            let valueDescription = (opportunities > 0) ? String(format: "%.0f (%d/%d)", value, count, opportunities) : "-"
+            return String(format: "%@: %@", name, valueDescription)
+        }
     }
         
     
     /// Total number of hands played.
-    /// Formula: Total Number of Hands Played
-    /// Function: cnt_hands
+    /// **Formula:** Total Number of Hands Played
+    /// **Function:** `cnt_hands`
     public var hands: Int { cnt_hands }
     
     /// Percentage of the time that a player voluntarily contributed money to the pot, given that he had a chance to do so.
-    /// Formula: Number of Times Player Put Money In Pot / (Number of Hands - Number of Walks)
-    /// Function: (cnt_vpip / (cnt_hands - cnt_walks)) * 100
+    /// **Formula:** Number of Times Player Put Money In Pot / (Number of Hands - Number of Walks)
+    /// **Function:** `(cnt_vpip / (cnt_hands - cnt_walks)) * 100`
     public var VPIP: Statistic
     {
         Statistic(
+            name: "VPIP",
             value: (Double(cnt_vpip) / Double(cnt_hands - cnt_walks)) * 100.0,
             count: cnt_vpip,
             opportunities: cnt_hands - cnt_walks
@@ -200,14 +208,114 @@ extension DetailedStatistics
     }
     
     /// Percentage of the time that a player put in any raise preflop, given that he had a chance to do so.
-    /// Formula: Number of Times Player Raised Preflop / (Number of Hands - Number of Walks)
-    /// Function: (cnt_pfr / cnt_pfr_opp) * 100
+    /// **Formula:** Number of Times Player Raised Preflop / (Number of Hands - Number of Walks)
+    /// **Function:** `(cnt_pfr / cnt_pfr_opp) * 100`
     public var PFR: Statistic
     {
         Statistic(
+        name: "PFR",
             value: (Double(cnt_pfr) / Double(cnt_pfr_opp)) * 100.0,
             count: cnt_pfr,
             opportunities: cnt_pfr_opp
         )
+    }
+    
+    /// Percentage of the time that a player opened the pot by raising from the cutoff, button, or small blind.
+    /// **Formula:** Number of Times Player Attempted to Steal Blinds / Number of Times Player Could Attempt to Steal Blinds.
+    /// **Function:** `(cnt_steal_att / cnt_steal_opp) * 100`
+    public var attemptToSteal: Statistic
+    {
+        Statistic(
+            name: "Attempt to Steal",
+            value: (Double(cnt_steal_att) / Double(cnt_steal_opp)) * 100,
+            count: cnt_steal_att,
+            opportunities: cnt_steal_opp
+        )
+    }
+    
+    /// Percentage of the time that a player folded when in a blind and facing an open raise from the cutoff, button, or small blind without any other players being involved.
+    /// **Formula:** Number of Times Player Folded Blind to a Steal / Number of Times Player Could Fold Blind to a Steal
+    /// **Function:** `(cnt_steal_def_action_fold / cnt_steal_def_opp) * 100`
+    public var foldToSteal: Statistic
+    {
+        Statistic(
+            name: "Fold to Steal",
+            value: (Double(cnt_steal_def_action_fold) / Double(cnt_steal_def_opp)) * 100,
+            count: cnt_steal_def_action_fold,
+            opportunities: cnt_steal_def_opp
+        )
+    }
+    
+    /// Percentage of the time that a player called from either blind when facing an open raise from the cutoff, button, or small blind.
+    /// **Formula:** Number of Times Player Called a Steal Attempt / Number of Times Player Could Call a Steal Attempt
+    /// **Function:** `(cnt_steal_def_action_call / cnt_steal_def_opp) * 100`
+    public var callSteal: Statistic
+    {
+        Statistic(
+            name: "Call Steal",
+            value: (Double(cnt_steal_def_action_call) / Double(cnt_steal_def_opp)) * 100,
+            count: cnt_steal_def_action_call,
+            opportunities: cnt_steal_def_opp
+        )
+    }
+    
+    /// Percentage of the time that a player 3Bet preflop when in a blind and facing an open raise from the cutoff, button, or small blind.
+    /// **Formula:** Number of Times Player 3Bet When Facing Steal / Number of Times Player Could 3Bet When Facing Steal
+    /// **Function:** `(cnt_steal_def_action_raise / cnt_steal_def_3bet_opp) * 100`
+    public var raiseSteal: Statistic
+    {
+        Statistic(
+            name: "Re-Steal",
+            value: (Double(cnt_steal_def_action_raise) / Double(cnt_steal_def_3bet_opp)) * 100,
+            count: cnt_steal_def_action_raise,
+            opportunities: cnt_steal_def_3bet_opp
+        )
+    }
+}
+
+
+
+extension DetailedStatistics
+{
+    
+    
+    
+    public var aligned: Aligned
+    { Aligned(statistics: self) }
+    
+    
+    public struct Aligned
+    {
+    
+        
+        let statistics: DetailedStatistics
+        
+    
+        /// Same as `PFR`, except it uses the same opportunities denominator as `VPIP`,
+        /// so PFR can be safely displayed as the subset of VPIP.
+        /// **Formula:** Number of Times Player Raised Preflop / (Number of Hands - Number of Walks)
+        /// **Function:** `(cnt_pfr / (cnt_hands - cnt_walks)) * 100`
+        public var PFR: Statistic
+        {
+            Statistic(
+                name: "PFR (aligned)",
+                value: (Double(statistics.cnt_pfr) / Double(statistics.cnt_hands - statistics.cnt_walks)) * 100.0,
+                count: statistics.cnt_pfr,
+                opportunities: statistics.cnt_hands - statistics.cnt_walks
+            )
+        }
+        
+        /// Same as `raiseSteal`, except it uses the same opportunities denominator as `foldToSteal`.
+        /// **Formula:** Number of Times Player 3Bet When Facing Steal / Number of Times Player Could 3Bet When Facing Steal
+        /// **Function:** `(cnt_steal_def_action_raise / cnt_steal_def_opp) * 100`
+        public var raiseSteal: Statistic
+        {
+            Statistic(
+                name: "Re-Steal (aligned)",
+                value: (Double(statistics.cnt_steal_def_action_raise) / Double(statistics.cnt_steal_def_opp)) * 100,
+                count: statistics.cnt_steal_def_action_raise,
+                opportunities: statistics.cnt_steal_def_opp
+            )
+        }
     }
 }